rename session.js to entities.js and associated commands to align with planned sample state; add servers to entities state
