+++ frontend/src/index.js
@@ -8,14 +8,14 @@ import './fonts.css';
 import App from './App';
 import configureStore from './store';
 import csrfFetch from './store/csrf';
-import * as sessionActions from './store/session';
+import * as entitiesActions from './store/entities';
 
 const store = configureStore();
 
 if (process.env.NODE_ENV !== "production") {
   window.store = store;
   window.csrfFetch = csrfFetch;
-  window.sessionActions = sessionActions;
+  window.entitiesActions = entitiesActions;
 }
 
 const Root = () => {
@@ -39,7 +39,7 @@ const renderApplication = () => {
 
 if (sessionStorage.getItem("currentUser") === null || 
   sessionStorage.getItem("X-CSRF-Token") === null) {
-  store.dispatch(sessionActions.restoreSession()).then(renderApplication);
+  store.dispatch(entitiesActions.restoreSession()).then(renderApplication);
 } else {
   renderApplication();
 }
